raw_kafka_feature_producer/requirements.txt
numpy
kafka-python